feature:适配fineio读写，支持long的pos

+++ swift-base/swift-base-io/src/main/java/com/fr/swift/cube/io/impl/fineio/input/ByteFineIoReader.java
@@ -27,6 +27,6 @@ public class ByteFineIoReader extends BaseFineIoReader<ByteBuf> implements ByteR
 
     @Override
     public byte get(long pos) {
-        return FineIOAccessor.INSTANCE.getByte(readFile, (int) pos);
+        return FineIOAccessor.INSTANCE.getByte(readFile, pos);
     }
 }
